Add Surface Code (#7)
This PR adds the `cudaq::qec::surface_code` namespace which includes the
`surface_code` type and the `stabilizer_grid` helper class. The
`stabilizer_grid` generates the 2d layout of the stabilizers, and which
data qubits each stabilizer has support on. Lastly it generates the
stabilizers and observables as vectors of `cudaq::spin_op` which the
`qec::code` expects.

---------

Co-authored-by: Ben Howe <141149032+bmhowe23@users.noreply.github.com>