Optimize `convert_vec_soft_to_tensor_hard` by 50x (#160)
This small PR optimizes one of our utility functions by >50x.

In the long term, we need to consider accessor methods for
`cudaqx::tensor` that don't incur the performance penalty of allocating
and deallocating vectors on every `.at()` access. As best as I can tell,
this is not supported very well by the underlying xtensor class.

Signed-off-by: Ben Howe <bhowe@nvidia.com>