Workflow updates (incl creating all_libs_release.yaml) (#159)
These workflow updates are needed to build the majority of the libraries
on the public runners. This is needed to ensure that the `__version__`
values return the right strings and right hashes.

The nv-qldpc-decoder version hash will still be from a private repo.

Note: this is a cherry-picked version of
https://github.com/NVIDIA/cudaqx/commit/e179eef5169cfc66ca2c21f4625e6e5af232548e,
which is on the releases/v0.3.0 branch now.

---------

Signed-off-by: Ben Howe <bhowe@nvidia.com>