Add Parity Check Matrix generation and utility functions (#147)
This PR introduces a CUDA-Q QEC Detector Error Model type (similar to
Stim's Detector Error Model but different).

A Detector Error Model can be generated from CUDA-Q circuits and then
fed to CUDA-Q QEC decoders as inputs so that the decoders know how to
decode the syndromes that come from those CUDA-Q circuits.

The main implementation and interfaces are done in C++. They are also
exposed to Python via pybind bindings.

In addition to the core functionality listed above, various PCM utility
functions (like column sorting, canonicalizing, combining, etc.) that
are used to create the DEM are also exposed to the user.

---------

Signed-off-by: Ben Howe <bhowe@nvidia.com>