Fix issue #258: added check and informative error message (#261)
This PR fixes issue #258 by adding a validation check on the dimensions
of the dem and the error rates and providing a clear, informative error
message when the condition is not met. This improves error handling and
helps users quickly identify the cause of the problem when the
construction is made by hand in a custom QEC example.

---------

Signed-off-by: Kevin Mato <kmato@nvidia.com>