Use 'tensor-network-decoder' in pip install instructions (#263)
The name`tensor_network_decoder` from the `.toml` file is normalized to
be `tensor-network-decoder` in the wheel's METADATA. Newer version of
`pip` is able to resolve this difference and `pip install
cudaq-qec[tensor_network_decoder]` will work correctly.

However, if you use an older version of `pip/setuptools`, you'll see an
error such as "cudaq-qec 0.4.0 does not provide the extra
'tensor_network_decoder'". A safer alternative is to use the exact name
as appeared in the METADATA (the normalized name with dash instead of
underscore) such that even earlier versions of `pip` will find the
package and install the dependencies as needed.

---------

Signed-off-by: Melody Ren <melodyr@nvidia.com>