Generalize `single_error_lut` to `multi_error_lut` (#281)
This retains the "single_error_lut" decoder and adds a "multi_error_lut"
decoder which is essentially the same as the "single_error_lut" but with
`lut_error_depth` > 1.

Also fixes #232 

Todo: docs need to come in a separate PR. We don't want them to go to
main right now since the released version doesn't have this decoder.
Tracked by #284

---------

Co-authored-by: melody-ren <melodyr@nvidia.com>