Fix address sanitizer issue (#323)
Worked this with @cketcham2333 ...

Prior to this change, if I compile `test_core` with address sanitization
support, I got these errors below. And after this PR, these errors go
away. It's unclear if these errors would impact us under normal
conditions or not because normally `Scalar` is a primitive type that
does not have a destructor.