Real-time decoding support (#333)
This PR adds support using our decoders in a real-time, online
environment, where the decoders are running on a low-latency connection
to a real QPU. The decoder corrections (calculated in the classical
computer) can be sent back to a live CUDA-Q program (running on the
quantum computer) during the quantum program's execution. Additionally,
one can run the decoder(s) in a simulated environment where real-time
decoding occurs in the simulation context (i.e. on the user's desktop).

---------

Signed-off-by: Ben Howe <bhowe@nvidia.com>
Co-authored-by: Chuck Ketcham <cketcham@nvidia.com>
Co-authored-by: Justin Lietz <jlietz@nvidia.com>
Co-authored-by: Kevin Mato <kmato@nvidia.com>
Co-authored-by: Alex McCaskey <amccaskey@nvidia.com>
Co-authored-by: Thien Nguyen <58006629+1tnguyen@users.noreply.github.com>
Co-authored-by: Sachin Pisal <spisal@nvidia.com>
Co-authored-by: melody-ren <melodyr@nvidia.com>