[core] Fix pre-existing extension point issue (#374)
This PR fixes a pre-existing issue with our extension points. Without
this change, when a new decoder plugin was added and cleanup_plugins was
called at the end of a program, the extension points would sometimes
crash. The solution is the following:

1. We must remove items from the registry to protect if someone were to
`dlclose` a plugin library while keeping the parent libcudaq-qec.so
library open.
2. Similar to #327, we change the static variables to pointers rather
than objects. This is actually the most important change.
3. We must use a mutex for protection against multi-threaded access to
the registry. It must be a recursive_mutex to allow decoders to include
other decoders (because the "get" method will try to lock multiple times
in the call stack).

Also, a bit unrelated, but I also changed one of our tests to use a
simulator that supports noise (stim) vs the default (which is often qpp
depending on the environment). This eliminates thousands of warning
messages that were introduced with
https://github.com/NVIDIA/cuda-quantum/pull/3113.

---------

Signed-off-by: Ben Howe <bhowe@nvidia.com>